Use a large enough data type for unregistered status codes.

+++ src/librusthttpserver/generate_status.py
@@ -135,17 +135,17 @@ pub enum Status {
                 out.write('\n    // %s\n' % entry)
 
         out.write('''
-    UnregisteredStatusCode(u8, ~str),
+    UnregisteredStatusCode(u16, ~str),
 }
 
 impl Status {
 
-    pub fn new(code: u8, reason: ~str) -> Status {
+    pub fn new(code: u16, reason: ~str) -> Status {
         UnregisteredStatusCode(code, reason)
     }
 
     /// Get the status code
-    pub fn code(&self) -> u8 {
+    pub fn code(&self) -> u16 {
         match *self {
 ''')
         for entry in data:
